requirements.txt
# ==============================================================================
# Pinned Dependencies for AI Agent
# Updated: 2025-06-18
# These specific versions resolve dependency conflicts and ensure stable builds
# ==============================================================================

# Core dependencies
numpy==1.24.3
langchain==0.3.25
langgraph==0.4.8
llama-index==0.10.0
llama-index-core==0.10.0
llama-index-readers-file==0.1.0
pydantic==2.5.0
typing-extensions==4.8.0

# AI/ML dependencies
torch==2.7.1
torchvision==0.22.1
transformers==4.36.0
sentence-transformers==2.2.2
scikit-learn==1.3.0

# Web framework and API
fastapi==0.104.1
gradio==5.25.2
uvicorn==0.24.0
aiofiles==23.2.1
aiohttp==3.9.1

# Data processing
pandas==2.1.4
openpyxl==3.1.2
pypdf==5.6.0
python-docx==1.1.0
python-pptx==0.6.23
unstructured==0.11.0

# Vision and media
opencv-python==4.11.0.86
Pillow==10.1.0
ffmpeg-python==0.2.0
librosa==0.10.1
moviepy==1.0.3
openai-whisper==20231117
pydub==0.25.1
yt-dlp==2023.12.30

# Database
psycopg2-binary==2.9.10
supabase==2.15.3

# Web scraping and data collection
beautifulsoup4==4.12.2
requests-html==0.10.0
selenium==4.15.2
wikipedia==1.4.0

# Resilience and monitoring
circuitbreaker==1.4.0
retry==0.9.2
tenacity==8.2.3
opentelemetry-api==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-sdk==1.21.0
prometheus-client==0.19.0

# Security
cryptography==41.0.8
python-jose==3.3.0

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# CLI and configuration
click==8.1.7
python-dotenv==1.0.0

# Utilities
colorama==0.4.6
ipython==8.18.1
matplotlib==3.8.2
python-chess==1.10.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3
requests==2.31.0
rich==13.7.0
seaborn==0.13.0
stockfish==3.28.0
tqdm==4.66.1
watchdog==3.0.0

# AI/LLM specific
anthropic==0.7.8
groq==0.4.2
langchain-experimental==0.0.49
langchain-groq==0.0.1
langchain-tavily==0.0.1
llama-index-embeddings-openai==0.1.0
llama-index-vector-stores-supabase==0.1.0
openai==1.3.7
tavily-python==0.3.1
autogen==0.2.0